requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
psycopg[binary,pool]==3.2.2
python-dotenv==1.0.1
sentence-transformers==3.0.1
numpy==1.26.4
orjson==3.10.7